chore: change Amount::total to i32 as it is unnecessary to support payment exceeding i32

+++ src/sdk/common/amount.rs
@@ -6,7 +6,7 @@ pub struct Amount {
   /// 订单总金额，单位为分。
   ///
   /// 示例值：100
-  pub total: i64,
+  pub total: i32,
   /// 货币类型
   ///
   /// CNY：人民币，境内商户号仅支持人民币。
